// src/generated/client/models/ProjectStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Lipsanen Project Management API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Project status
 * @export
 */
export const ProjectStatus = {
    Initiation: 'INITIATION',
    Planning: 'PLANNING',
    Design: 'DESIGN',
    Procurement: 'PROCUREMENT',
    Construction: 'CONSTRUCTION',
    Inspection: 'INSPECTION',
    Completion: 'COMPLETION'
} as const;
export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export function ProjectStatusFromJSON(json: any): ProjectStatus {
    return ProjectStatusFromJSONTyped(json, false);
}

export function ProjectStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectStatus {
    return json as ProjectStatus;
}

export function ProjectStatusToJSON(value?: ProjectStatus | null): any {
    return value as any;
}